Refactor tickrate changer
